requirements.txt
streamlit==1.32.0
python-docx==1.1.0
pdfplumber==0.10.3
nltk==3.8.1
scikit-learn==1.4.0
pandas==2.2.0 